a2a_agent/requirements.txt
a2a-sdk
google-adk
google-genai
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-firestore
google-cloud-storage
pandas
numpy
matplotlib
seaborn
immutabledict
regex
sqlglot
praw
uvicorn
starlette
gunicorn
python-dotenv
asyncclick
requests
fastapi
python-multipart